chore: move custom mm preprocessor to new file
